requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
requests==2.25.1
python-slugify==6.1.1
